model/crop_disease_detection/requirements.txt
Flask
flask-cors
gunicorn
numpy
pillow
tflite-runtime
